Intermediate update after fixing the display bug in the Debug_Data_Display_Controller (The display was not tied to the camera movement and flickered when moving the Avatar_Component)

+++ Tunnelour/Tunnelour/src/Background_Controller.cc
@@ -69,7 +69,7 @@ void Background_Controller::Run() {
     int resised_tile_size = static_cast<int>(base_tile_size * m_game_settings->GetTileMultiplicationFactor());
 
     int number_of_y_tiles = 0;
-    std::div_t div_y_result =  div(m_game_settings->GetResolution().y, resised_tile_size);
+    std::div_t div_y_result =  div(static_cast<int>(m_game_settings->GetResolution().y), resised_tile_size);
     if (div_y_result.rem != 0) {
       number_of_y_tiles = div_y_result.quot + 1;
     } else {
@@ -78,7 +78,7 @@ void Background_Controller::Run() {
 
     //Calculate number of x tiles
     int number_of_x_tiles = 0;
-    std::div_t div_x_result = div(m_game_settings->GetResolution().x, resised_tile_size);
+    std::div_t div_x_result = div(static_cast<int>(m_game_settings->GetResolution().x), resised_tile_size);
     if (div_x_result.rem != 0) {
       number_of_x_tiles = div_x_result.quot + 1;
     } else {
@@ -382,7 +382,7 @@ Tunnelour::Tile_Bitmap* Background_Controller::Create_Tile(int base_tile_size, i
                                                       static_cast<float>(middleground_line.top_left_y));
 
   //tile->SetSize(new D3DXVECTOR2(middleground_line.tile_size_x * 4, middleground_line.tile_size_x * 4));
-  tile->SetSize(new D3DXVECTOR2(resised_tile_size, resised_tile_size));
+  tile->SetSize(new D3DXVECTOR2(static_cast<float>(resised_tile_size), static_cast<float>(resised_tile_size)));
   return tile;
 }
